Dockerhub (#3)
* feat(loans): complete microservice and refactor codebase

* add docker-compose.yml for accounts, cards, and loans microservices

* refactor docker setup: single docker-compose file using remote Docker Hub images for accounts, cards, and loans microservices deployment

* feat(config): add config server

* feat: integrate microservices with Consul discovery

+++ accounts/src/main/java/theo_viti/com/example/accounts/model/Accounts.java
@@ -13,14 +13,18 @@ import java.time.LocalDate;
 public class Accounts {
 
       @Column(name = "customer_id")
-      private int customerId;
+      private Long customerId;
+
       @Column(name="account_number")
       @Id
-      private long accountNumber;
+      private Long accountNumber;
+
       @Column(name="account_type")
       private String accountType;
+
       @Column(name = "branch_address")
       private String branchAddress;
+
       @Column(name = "create_dt")
       private LocalDate createDt;
